+++ MUMENT/MUMENT/Sources/Scenes/Home/Model/DefaultNavigationBarModel.swift
@@ -4,12 +4,12 @@
 //
 //  Created by 김지민 on 2022/07/15.
 //
-
-import UIKit
-
-struct DefaultNavigationBarModel {
-    let title: String
-}
+//
+//import UIKit
+//
+//struct DefaultNavigationBarModel {
+//    let title: String
+//}
 
 //// MARK: - Extensions
 //extension DefaultNavigationBarModel {
